requirements.txt
Django>=5.2.7
djangorestframework>=3.15.0
redis>=5.0.0
django-redis>=5.4.0
celery>=5.4.0
django-celery-results>=2.5.1
django-celery-beat>=2.5.0
python-dotenv>=1.0.0
docker>=6.1.0
PyMySQL>=1.1.0
pytest>=8.0.0
pytest-django>=4.7.0
coverage>=7.0.0
requests>=2.31.0
Faker>=23.0.0